기초 도메인 설계 및 간이 기초구역 검색 테스트 코드 작성

+++ src/main/java/com/yapp/fmz/domain/Location.java
@@ -7,6 +7,6 @@ import javax.persistence.Embeddable;
 @Embeddable
 @Getter
 public class Location {
-    Float lat;
-    Float lng;
+    Double lat;
+    Double lng;
 }
